backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
boto3==1.34.140
botocore==1.34.140
celery==5.4.0
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
cryptography==42.0.8
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-storages==1.14.3
django-timezone-field==7.0
jmespath==1.0.1
kombu==5.3.7
prompt_toolkit==3.0.47
pycparser==2.22
PyMySQL==1.1.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
django-environ==0.8.1
djangorestframework==3.14.0
drf-yasg==1.21.4
Pillow==10.0.0
django-cors-headers==3.13.0
httpx==0.24.0
python-dotenv==0.21.0
django-prometheus
django-redis==5.2.0
logstash-formatter
elasticsearch
